Move ValHamiltionain from BloqadeKrylov to BloqadeExpr.Lowlevel
