+++ lib/BloqadeExpr/src/Lowlevel/Lowlevel.jl
@@ -10,11 +10,12 @@ using Adapt
 using LaTeXStrings
 using LinearAlgebra
 
-export Hamiltonian, StepHamiltonian
+export Hamiltonian, StepHamiltonian, ValHamiltonian
 export ThreadedMatrix
 export set_backend
 export storage_size, to_matrix
 export precision_type, highest_type
+export ValH, get_f # convert StepHamiltonian to ValHamiltonian
 
 include("types.jl")
 include("printings.jl")
